+++ src/assignments_reviews/lab002/Question4_Females.java
@@ -2,5 +2,5 @@ package assignments_reviews.lab002;
 
 public class Question4_Females {
 
-    int counter;
+    double counter;
 }
