+++ src/assignments_reviews/lab002/Question4_Males.java
@@ -2,7 +2,7 @@ package assignments_reviews.lab002;
 
 public class Question4_Males {
 
-    int counter;
+    double counter;
 
 
 }
